week_4/requirements.txt
# Core LangChain and LangGraph
langchain>=0.3.0
langgraph>=0.4.8
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0

# FastAPI and server
fastapi>=0.115.0
uvicorn>=0.34.0
python-multipart>=0.0.9

# Streamlit UI
streamlit>=1.39.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0

# Development tools
watchdog>=6.0.0
langgraph-cli[inmem]>=0.3.0